+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] <!-- release-date -->
 
+### Added
+
+- Implement "find institutions" functionality on the institutions service client.
+
 ### Changed
 
 - In hosted_license_provider_lifecycle test, re-enable assertation of retrieved method icon, as the issue was fixed upstream.